refactor: update vite to latest version
- simplify path aliases and refactor all files
- fix ui bug with nav buttons not updating their styles properly
- #357
