+++ frontend/package.json
@@ -32,9 +32,11 @@
     "@mui/x-charts": "^7.27.0",
     "@typescript-eslint/eslint-plugin": "^8.11.0",
     "@typescript-eslint/parser": "^8.11.0",
+    "clsx": "^2.1.1",
     "file-saver": "^2.0.5",
     "framer-motion": "^11.11.11",
     "http-status-codes": "^2.3.0",
+    "ogl": "^1.0.11",
     "papaparse": "^5.4.1",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.5",
@@ -42,21 +44,22 @@
     "react-router-dom": "^6.26.2",
     "styled-components": "^6.1.13",
     "uuid": "^10.0.0",
-    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz",
-    "ogl": "^1.0.11"
+    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
+    "@tailwindcss/vite": "^4.0.17",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@types/file-saver": "^2.0.7",
+    "@types/node": "^22.13.14",
     "@types/papaparse": "^5.3.14",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.0",
     "@types/supertest": "^6.0.2",
     "@types/uuid": "^10.0.0",
-    "@vitejs/plugin-react": "^4.3.1",
+    "@vitejs/plugin-react": "^4.3.4",
     "@vitest/coverage-v8": "^2.1.4",
     "autoprefixer": "^10.4.20",
     "eslint": "^9.13.0",
@@ -69,10 +72,10 @@
     "postcss": "^8.4.47",
     "prettier": "^3.5.2",
     "supertest": "^7.0.0",
-    "tailwindcss": "^3.4.13",
+    "tailwindcss": "^4.0.17",
     "typescript": "^5.6.2",
     "typescript-eslint": "^8.11.0",
-    "vite": "^5.4.1",
+    "vite": "^6.2.3",
     "vite-tsconfig-paths": "^5.1.0",
     "vitest": "^2.1.4"
   }
