+++ frontend/src/context/index.ts
@@ -2,3 +2,5 @@ export * from "./CourseProvider";
 export * from "./AssignmentProvider";
 export * from "./RubricContext";
 export * from "./DialogContext";
+export * from "./SettingsContext";
+export * from "./GradingContext";
