+++ frontend/src/features/grading/projectGradingComponents/GroupFeedback.tsx
@@ -1,4 +1,4 @@
-import { useGradingContext } from "../../../context/GradingContext.tsx";
+import { useGradingContext } from "@/context";
 
 export function GroupFeedback() {
   const { gradedSubmissionCache, updateGroupComment } = useGradingContext();
