+++ frontend/src/features/templatesPage/TempatesMain.tsx
@@ -3,15 +3,16 @@
  */
 
 import { ReactElement, useEffect, useState } from "react";
-import { ChoiceDialog, Dialog, MainPageTemplate } from "@components";
+import { ChoiceDialog, Dialog, MainPageTemplate } from "@/components";
 import { TemplateProvider, useTemplatesContext } from "./TemplateContext.tsx";
 import TemplatesWindow from "./TemplatesWindow.tsx";
 import TemplateSearch from "./TemplateSearch.tsx";
 import AddTemplateTag from "./AddTemplateTag.tsx";
-import { TemplateBuilder } from "src/features/templatesPage/TemplateBuilder.tsx";
+import { TemplateBuilder } from "@/features/templatesPage/TemplateBuilder.tsx";
 import { Template } from "palette-types";
-import { useChoiceDialog } from "../../context/DialogContext.tsx";
+import { useChoiceDialog } from "@/context/DialogContext.tsx";
 import TemplateCharts from "./TemplateCharts.tsx";
+
 export default function TemplatesMain(): ReactElement {
   return (
     <TemplateProvider>
@@ -29,7 +30,6 @@ function TemplatesMainContent(): ReactElement {
     setShowSuggestions,
     handleCreateTemplate,
     handleSubmitNewTemplate,
-    handleQuickStart,
     setIsNewTemplate,
     setShowBulkActions,
     editingTemplate,
@@ -149,14 +149,6 @@ function TemplatesMainContent(): ReactElement {
           >
             Create Template
           </button>
-          {templates?.length === 0 && (
-            <button
-              onClick={() => void handleQuickStart()}
-              className="bg-blue-500 text-white font-bold mb-6 rounded-lg py-2 px-4 mr-4 hover:bg-blue-700 hover:scale-105 focus:outline-none focus:ring-2 focus:ring-blue-500"
-            >
-              Quick Start
-            </button>
-          )}
         </div>
 
         <Dialog
