+++ frontend/src/hooks/index.ts
@@ -5,3 +5,5 @@
 export * from "./useFetch";
 export * from "./useLocalStorage";
 export * from "./useUpdateLogger";
+export * from "./useTemplate";
+export * from "./useRubricBuilder";
