+++ frontend/src/utils/index.ts
@@ -5,5 +5,7 @@
 export * from "./time.ts";
 export * from "./rubricFactory";
 export * from "./calculateMaxPoints";
-export * from "../features/rubricBuilder/csv/CSVParser.ts";
+export * from "../features/rubricBuilder/csv/CSVParser";
 export * from "./tagFactory";
+export * from "./SubmissionUtils";
+export * from "./templateFactory.ts";
