+++ frontend/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
+    "target": "esnext",
+    "module": "esnext",
     "jsx": "react-jsx",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "baseUrl": "./",
     "rootDir": "../",
     "strict": true,
@@ -12,19 +13,51 @@
     "forceConsistentCasingInFileNames": true,
     "sourceMap": true,
     "allowImportingTsExtensions": true,
-    "lib": ["dom", "ESNext"],
-    "types": ["node", "react", "vite/client", "vitest"],
+    "lib": [
+      "dom",
+      "ESNext"
+    ],
+    "types": [
+      "node",
+      "react",
+      "vite/client",
+      "vitest"
+    ],
     "resolveJsonModule": true,
     "paths": {
-      "@utils": ["src/utils"],
-      "@components": ["src/components"],
-      "@features": ["src/features"],
-      "@context": ["src/context"],
-      "@hooks": ["src/hooks"],
-      "@resources": ["src/resources"],
-      "@local_types": ["src/types"]
+      "@/*": [
+        "./src/*"
+      ],
+      "@utils": [
+        "src/utils"
+      ],
+      "@components": [
+        "src/components"
+      ],
+      "@features": [
+        "src/features"
+      ],
+      "@context": [
+        "src/context"
+      ],
+      "@hooks": [
+        "src/hooks"
+      ],
+      "@resources": [
+        "src/resources"
+      ],
+      "@local_types": [
+        "src/types"
+      ]
     }
   },
-  "include": ["src", "vite.config.ts", "**/*.test.ts"],
-  "exclude": ["node_modules", "dist"]
+  "include": [
+    "src",
+    "vite.config.ts",
+    "**/*.test.ts"
+  ],
+  "exclude": [
+    "node_modules",
+    "dist"
+  ]
 }
