+++ frontend/vite.config.ts
@@ -1,11 +1,16 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import tsconfigPaths from "vite-tsconfig-paths";
+import path from "path";
+import tailwindcss from "@tailwindcss/vite";
 import { configDefaults } from "vitest/config";
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), tsconfigPaths()],
+  plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "/src"),
+    },
+  },
   base: "./",
   build: {
     outDir: "../backend/dist/frontend",
