+++ package.json
@@ -23,5 +23,7 @@
   },
   "devDependencies": {
     "concurrently": "^9.0.1"
+  },
+  "dependencies": {
   }
 }
